+++ apis/getambassador.io/v1beta2/common.go
@@ -1,6 +1,7 @@
 package v1
 
 import (
+	"bytes"
 	"encoding/json"
 	"net/http"
 	"text/template"
@@ -81,5 +82,70 @@ func (hf *HeaderFieldTemplate) Validate() error {
 		return errors.Wrapf(err, "parsing template for header %q", hf.Name)
 	}
 	hf.Template = tmpl
+	return nil
+}
+
+type ErrorResponse struct {
+	Realm string `json:"realm"`
+
+	ContentType string                `json:"contentType"`
+	Headers     []HeaderFieldTemplate `json:"headers"`
+
+	RawBodyTemplate string             `json:"bodyTemplate"`
+	BodyTemplate    *template.Template `json:"-"`
+}
+
+func (er *ErrorResponse) Validate(qname string) error {
+	// Handle deprecated .ContentType
+	if er.ContentType != "" {
+		er.Headers = append(er.Headers, HeaderFieldTemplate{
+			Name:  "Content-Type",
+			Value: er.ContentType,
+		})
+	}
+
+	// Fill defaults
+	if er.Realm == "" {
+		er.Realm = qname
+	}
+	if len(er.Headers) == 0 {
+		er.Headers = append(er.Headers, HeaderFieldTemplate{
+			Name:  "Content-Type",
+			Value: "application/json",
+		})
+	}
+	if er.RawBodyTemplate == "" {
+		er.RawBodyTemplate = `{{ . | json "" }}`
+	}
+
+	// Parse+validate the header-field templates
+	for i := range er.Headers {
+		hf := &(er.Headers[i])
+		if err := hf.Validate(); err != nil {
+			return errors.Wrap(err, "headers")
+		}
+	}
+	// Parse+validate the bodyTemplate
+	tmpl, err := template.
+		New("bodyTemplate").
+		Funcs(template.FuncMap{
+			"json": func(prefix string, data interface{}) (string, error) {
+				nonIdentedJSON, err := json.Marshal(data)
+				if err != nil {
+					return "", err
+				}
+				var indentedJSON bytes.Buffer
+				if err := json.Indent(&indentedJSON, nonIdentedJSON, prefix, "\t"); err != nil {
+					return "", err
+				}
+				return indentedJSON.String(), nil
+			},
+		}).
+		Parse(er.RawBodyTemplate)
+	if err != nil {
+		return errors.Wrap(err, "parsing template for bodyTemplate")
+	}
+	er.BodyTemplate = tmpl
+
 	return nil
 }